Desktop/Profile_project/Hiring_AI/backend/requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
python-multipart==0.0.9

# Database
supabase==2.3.4
asyncpg>=0.30.0
pgvector==0.2.4

# AI / OpenAI
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.8
tiktoken>=0.7.0

# PDF Parsing
pymupdf>=1.24.0

# Async / Task Queue
celery==5.3.6
redis==5.0.1

# AWS
boto3==1.34.34
botocore==1.34.34

# Auth & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Config
pydantic-settings==2.2.1
python-dotenv==1.0.1

# HTTP client
httpx==0.26.0

# Data Processing
numpy>=2.1.0
scipy>=1.14.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.26.0
